Fix printf format specifiers in recv demo
recv.cpp printed msg.id (a uint) and msg.size (a __uint128_t) both with %d,
which is the wrong type and the wrong size. With -Wall g++ flagged the size
print: 'format %d expects int, but argument has type __int128 unsigned'.

Use %u for the unsigned id and %llu with an explicit (unsigned long long)
cast for size (there is no portable printf specifier for 128-bit integers;
the demo value, sizeof(int), fits comfortably). Output is unchanged:
rec 70 / rec 4.

+++ recv.cpp
@@ -10,8 +10,8 @@ int main(int argc, char const *argv[])
     while (!p.receive(&msg))
         ;
 
-    printf("rec %d\n", msg.id);
-    printf("rec %d\n", msg.size);
+    printf("rec %u\n", msg.id);
+    printf("rec %llu\n", (unsigned long long)msg.size);
 
     return 0;
 }
